requirements.txt
djangorestframework
python-decouple
psycopg2-binary
dj-database-url